other small fixes
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@2737 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/kvirc/ui/kvi_channel.cpp
@@ -1841,7 +1841,7 @@ void KviChannel::pasteLastLog()
 		iLogDay = szLogDate.section('.',2,2).toInt();
 
 		// Check log validity
-		int iInterval = -KVI_OPTION_UINT(KviOption_uintDaysIntervalToPasteOnChannelJoin);
+		int iInterval = -(int)KVI_OPTION_UINT(KviOption_uintDaysIntervalToPasteOnChannelJoin);
 		QDate logDate(iLogYear,iLogMonth,iLogDay);
 		QDate checkDate = date.addDays(iInterval);
